rename and simplfy test vertex
