+++ gfe_integration_tests/sdram_edge_tests/test_c_sdram_with_recording/test_sdram_recording.py
@@ -12,11 +12,10 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-#from fec_integration_tests.interface.interface_functions.\
-#    simple_test_vertex import SimpleTestVertex
 from pacman_test_objects import SimpleTestVertex
 from gfe_integration_tests.sdram_edge_tests import (
     test_c_sdram_with_recording)
+from gfe_integration_tests.sdram_test_vertex import SdramTestVertex
 from gfe_integration_tests.sdram_edge_tests.\
     test_c_sdram_with_recording import (
         SDRAMSplitterExternal)
@@ -30,10 +29,10 @@ class TestSDRAMEdgeWithRecording(BaseTestCase):
 
     def setup(self):
         sim.setup(model_binary_module=test_c_sdram_with_recording)
-        vertex_1 = SimpleTestVertex(2) #, fixed_sdram_value=20)
+        vertex_1 = SdramTestVertex(2, fixed_sdram_value=20)
         vertex_1.splitter = SDRAMSplitterExternal(
             ConstantSDRAMMachinePartition)
-        vertex_2 = SimpleTestVertex(2) #, fixed_sdram_value=20)
+        vertex_2 = SdramTestVertex(2, fixed_sdram_value=20)
         vertex_2.splitter = SDRAMSplitterExternal(
             ConstantSDRAMMachinePartition)
         sim.add_vertex_instance(vertex_1)
