+++ gfe_integration_tests/sdram_edge_tests/test_constant_internal/test_sdram_internal_edges.py
@@ -12,7 +12,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from gfe_integration_tests.simple_test_vertex import SimpleTestVertex
+from gfe_integration_tests.sdram_test_vertex import SdramTestVertex
 from gfe_integration_tests.sdram_edge_tests.common import (
     SDRAMSplitterInternal)
 from gfe_integration_tests.sdram_edge_tests import common
@@ -25,7 +25,7 @@ class TestConstantSDRAMEdgeInsideOneAppVert(BaseTestCase):
 
     def setup(self):
         sim.setup(model_binary_module=common)
-        vertex_1 = SimpleTestVertex(10, fixed_sdram_value=20)
+        vertex_1 = SdramTestVertex(10, fixed_sdram_value=20)
         vertex_1.splitter = SDRAMSplitterInternal(
             ConstantSDRAMMachinePartition)
         sim.add_vertex_instance(vertex_1)
