+++ gfe_integration_tests/sdram_edge_tests/test_multi_dest_internal/test_sdram_internal_edges.py
@@ -12,7 +12,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from gfe_integration_tests.simple_test_vertex import SimpleTestVertex
+from gfe_integration_tests.sdram_test_vertex import SdramTestVertex
 from gfe_integration_tests.sdram_edge_tests.\
     test_multi_dest_internal import SDRAMSplitter
 from pacman.model.graphs.machine import (
@@ -26,7 +26,7 @@ class TestMultiDestSDRAMEdgeInsideOneAppVert(BaseTestCase):
 
     def setup(self):
         sim.setup(model_binary_module=common)
-        vertex_1 = SimpleTestVertex(12, fixed_sdram_value=20)
+        vertex_1 = SdramTestVertex(12, fixed_sdram_value=20)
         vertex_1.splitter = SDRAMSplitter(
             DestinationSegmentedSDRAMMachinePartition)
         sim.add_vertex_instance(vertex_1)
