+++ gfe_integration_tests/sdram_edge_tests/test_multi_src_external/test_sdram_external_edges.py
@@ -12,7 +12,7 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
-from gfe_integration_tests.simple_test_vertex import SimpleTestVertex
+from gfe_integration_tests.sdram_test_vertex import SdramTestVertex
 from pacman.model.graphs.application import ApplicationEdge
 from pacman.model.graphs.machine import SourceSegmentedSDRAMMachinePartition
 from gfe_integration_tests.sdram_edge_tests import common
@@ -28,9 +28,9 @@ class TestMultiSrcSDRAMEdgeOverTwoAppVert(BaseTestCase):
 
     def setup(self):
         sim.setup(model_binary_module=common)
-        vertex_1 = SimpleTestVertex(12, fixed_sdram_value=20)
+        vertex_1 = SdramTestVertex(12, fixed_sdram_value=20)
         vertex_1.splitter = BasicSDRAMSplitter()
-        vertex_2 = SimpleTestVertex(12, fixed_sdram_value=20)
+        vertex_2 = SdramTestVertex(12, fixed_sdram_value=20)
         vertex_2.splitter = SDRAMSplitter(
             SourceSegmentedSDRAMMachinePartition, vertex_1.splitter)
         sim.add_vertex_instance(vertex_1)
